[#24]refactor: 토큰 관련 로직을 /lib 로 분리
